Refactor demo and add bootstrap
